online_movie_bd/requirements.txt
psycopg2-binary
streamlit
pandas
bcrypt